Add missing topics and related cleanup (BL-9496)
