--- src/model/CloseVocabularies.ts
@@ -1,36 +1,0 @@
-export const kTopicList = [
-    "Agriculture",
-    "Animal Stories",
-    "Business",
-    "Dictionary",
-    "Environment",
-    "Primer",
-    "Math",
-    "Culture",
-    "Science",
-    "Story Book",
-    "Traditional Story",
-    "Health",
-    "Personal Development",
-    "Spiritual",
-];
-
-interface IStringMap {
-    [id: string]: string;
-}
-export const kNameToL10NKey: IStringMap = {
-    "topic:Agriculture": "topic.Agriculture",
-    "topic:Animal Stories": "topic.Animal Stories",
-    "topic:Business": "topic.Business",
-    "topic:Dictionary": "topic.Dictionary",
-    "topic:Environment": "topic.Environment",
-    "topic:Primer": "topic.Primer",
-    "topic:Math": "topic.Math",
-    "topic:Culture": "topic.Culture",
-    "topic:Science": "topic.Science",
-    "topic:Story Book": "topic.Story Book",
-    "topic:Traditional Story": "topic.Traditional Story",
-    "topic:Health": "topic.Health",
-    "topic:Personal Development": "topic.Personal Development",
-    "topic:Spiritual": "topic.Spiritual",
-};
