+++ src/model/ClosedVocabularies.ts
@@ -2,15 +2,43 @@ export const kTopicList = [
     "Agriculture",
     "Animal Stories",
     "Business",
+    "Community Living",
+    "Culture",
     "Dictionary",
     "Environment",
-    "Primer",
+    "Fiction",
+    "Health",
+    "How To",
     "Math",
-    "Culture",
+    "Non Fiction",
+    "Personal Development",
+    "Primer",
     "Science",
+    "Spiritual",
     "Story Book",
     "Traditional Story",
-    "Health",
-    "Personal Development",
-    "Spiritual",
 ];
+
+interface IStringMap {
+    [id: string]: string;
+}
+export const kNameToL10NKey: IStringMap = {
+    "topic:Agriculture": "topic.Agriculture",
+    "topic:Animal Stories": "topic.Animal Stories",
+    "topic:Business": "topic.Business",
+    "topic:Community Living": "topic.Community Living",
+    "topic:Culture": "topic.Culture",
+    "topic:Dictionary": "topic.Dictionary",
+    "topic:Environment": "topic.Environment",
+    "topic:Fiction": "topic.Fiction",
+    "topic:Health": "topic.Health",
+    "topic:How To": "topic.How To",
+    "topic:Math": "topic.Math",
+    "topic:Non Fiction": "topic.Non Fiction",
+    "topic:Personal Development": "topic.Personal Development",
+    "topic:Primer": "topic.Primer",
+    "topic:Science": "topic.Science",
+    "topic:Spiritual": "topic.Spiritual",
+    "topic:Story Book": "topic.Story Book",
+    "topic:Traditional Story": "topic.Traditional Story",
+};
